Setup

+++ db/migrations/20171029133347_more-school-metrics.js
@@ -218,7 +218,7 @@ exports.up = function(knex, Promise) {
       table.decimal('evidence_based_reading_writing_mean_score', 4, 1);
       table.decimal('math_mean_score', 4, 1);
       table.decimal('overall_mean_score', 4, 1);
-      table.integer('participation_rate', 4, 3);
+      table.decimal('participation_rate', 4, 3);
       table.integer('valid_scores_prev_year');
       table.decimal('evidence_based_reading_writing_mean_score_prev_year', 4, 1);
       table.decimal('math_mean_score_prev_year', 4, 1);
